sydyn/requirements.txt
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
requests>=2.31.0
anthropic>=0.42.0
python-dotenv>=1.0.1
fastapi>=0.109.0
uvicorn>=0.27.0
sse-starlette>=1.8.0
